# api/openapi/swagger.yaml
openapi: 3.1.0
info:
  title: Temperature Log API
  version: 0.1.0
servers:
- url: /
paths:
  /temperatures:
    get:
      operationId: listTemperatureLogs
      parameters:
      - explode: true
        in: query
        name: probe_id
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: start_at
        required: false
        schema:
          format: date-time
          type: string
        style: form
      - explode: true
        in: query
        name: end_at
        required: false
        schema:
          format: date-time
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/TemperatureLog"
                type: array
          description: OK
      summary: Search temperature readings
    post:
      operationId: createTemperatureLog
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemperatureCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureLog"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request
      summary: Register a temperature reading
  /temperatures/bulk:
    post:
      operationId: createTemperatureLogs
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemperatureCreateRequests"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/TemperatureLog"
                type: array
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request
      summary: Register multiple temperature readings
  /temperatures/latest/{probe_id}:
    get:
      operationId: getLatestTemperatureLog
      parameters:
      - explode: false
        in: path
        name: probe_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureLog"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
      summary: Get the latest temperature reading
  /probes:
    get:
      operationId: listProbes
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Probe"
                type: array
          description: OK
      summary: List probes
  /probes/{probe_id}:
    delete:
      operationId: deleteProbe
      parameters:
      - explode: false
        in: path
        name: probe_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
      summary: Delete a probe
  /experiments:
    get:
      operationId: listExperiments
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Experiment"
                type: array
          description: OK
      summary: List experiments
    post:
      operationId: createExperiment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperimentCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request
      summary: Create a new experiment
  /experiments/current:
    get:
      operationId: getCurrentExperiment
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
      summary: Get the current experiment
  /experiments/{experiment_id}:
    get:
      operationId: getExperiment
      parameters:
      - explode: false
        in: path
        name: experiment_id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
      summary: Get an experiment
    delete:
      operationId: deleteExperiment
      parameters:
      - explode: false
        in: path
        name: experiment_id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid state
      summary: Archive an experiment
  /experiments/{experiment_id}/start:
    post:
      operationId: startExperiment
      parameters:
      - explode: false
        in: path
        name: experiment_id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid state
      summary: Start an experiment
  /experiments/{experiment_id}/end:
    post:
      operationId: endExperiment
      parameters:
      - explode: false
        in: path
        name: experiment_id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid state
      summary: End an experiment
  /experiments/{experiment_id}/complete:
    post:
      operationId: completeExperiment
      parameters:
      - explode: false
        in: path
        name: experiment_id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid state
      summary: Complete an experiment
  /experiments/{experiment_id}/reopen:
    post:
      operationId: reopenExperiment
      parameters:
      - explode: false
        in: path
        name: experiment_id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid state
      summary: Reopen an experiment
  /experiments/{experiment_id}/probes:
    get:
      operationId: listExperimentProbes
      parameters:
      - explode: false
        in: path
        name: experiment_id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ExperimentProbe"
                type: array
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
      summary: List probes assigned to an experiment
    post:
      operationId: addExperimentProbe
      parameters:
      - explode: false
        in: path
        name: experiment_id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperimentProbeCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentProbe"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
      summary: Assign a probe to an experiment
components:
  schemas:
    TemperatureLog:
      example:
        recorded_at: 2000-01-23T04:56:07.000+00:00
        probe_id: probe_id
        temperature: 6.0274563
        id: 0
      properties:
        id:
          format: int64
          title: id
          type: integer
        probe_id:
          title: probe_id
          type: string
        recorded_at:
          format: date-time
          title: recorded_at
          type: string
        temperature:
          format: float
          title: temperature
          type: number
      required:
      - id
      - probe_id
      - recorded_at
      - temperature
      title: TemperatureLog
    TemperatureCreateRequest:
      example:
        recorded_at: 2000-01-23T04:56:07.000+00:00
        probe_id: probe_id
        temperature: 0.8008282
      properties:
        probe_id:
          title: probe_id
          type: string
        recorded_at:
          format: date-time
          title: recorded_at
          type: string
        temperature:
          format: float
          title: temperature
          type: number
      required:
      - probe_id
      - temperature
      title: TemperatureCreateRequest
    TemperatureCreateRequests:
      items:
        $ref: "#/components/schemas/TemperatureCreateRequest"
      title: TemperatureCreateRequests
      type: array
    Experiment:
      example:
        ended_at: 2000-01-23T04:56:07.000+00:00
        completed_at: 2000-01-23T04:56:07.000+00:00
        archived_at: 2000-01-23T04:56:07.000+00:00
        created_at: 2000-01-23T04:56:07.000+00:00
        name: Experiment 001
        description: cooling test
        updated_at: 2000-01-23T04:56:07.000+00:00
        started_at: 2000-01-23T04:56:07.000+00:00
        status: completed
        id: 0
      properties:
        id:
          format: int64
          title: id
          type: integer
        name:
          title: name
          type: string
        description:
          title: description
          type: string
        status:
          enum:
          - planned
          - running
          - completed
          - archived
          - finished
          title: status
          type: string
        started_at:
          format: date-time
          title: started_at
          type: string
        ended_at:
          format: date-time
          title: ended_at
          type: string
        completed_at:
          format: date-time
          title: completed_at
          type: string
        archived_at:
          format: date-time
          title: archived_at
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      required:
      - id
      - name
      - status
      - created_at
      - updated_at
      title: Experiment
    ExperimentCreateRequest:
      example:
        name: Experiment 001
        description: cooling test
      properties:
        name:
          title: name
          type: string
        description:
          title: description
          type: string
      required:
      - name
      title: ExperimentCreateRequest
    ExperimentProbe:
      example:
        valid_to: 2000-01-23T04:56:07.000+00:00
        created_at: 2000-01-23T04:56:07.000+00:00
        probe_id: probe-a
        experiment_id: 0
        updated_at: 2000-01-23T04:56:07.000+00:00
        role: target
        valid_from: 2000-01-23T04:56:07.000+00:00
        id: 0
      properties:
        id:
          format: int64
          title: id
          type: integer
        experiment_id:
          format: int64
          title: experiment_id
          type: integer
        probe_id:
          title: probe_id
          type: string
        role:
          title: role
          type: string
        valid_from:
          format: date-time
          title: valid_from
          type: string
        valid_to:
          format: date-time
          title: valid_to
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      required:
      - id
      - experiment_id
      - probe_id
      - role
      - created_at
      - updated_at
      title: ExperimentProbe
    ExperimentProbeCreateRequest:
      example:
        probe_id: probe-a
        role: target
        valid_from: 2000-01-23T04:56:07.000+00:00
        valid_to: 2000-01-23T04:56:07.000+00:00
      properties:
        probe_id:
          title: probe_id
          type: string
        role:
          title: role
          type: string
        valid_from:
          format: date-time
          title: valid_from
          type: string
        valid_to:
          format: date-time
          title: valid_to
          type: string
      required:
      - probe_id
      - role
      title: ExperimentProbeCreateRequest
    Probe:
      example:
        created_at: 2000-01-23T04:56:07.000+00:00
        probe_id: probe-a
        deleted_at: 2000-01-23T04:56:07.000+00:00
        updated_at: 2000-01-23T04:56:07.000+00:00
        id: 0
      properties:
        id:
          format: int64
          title: id
          type: integer
        probe_id:
          title: probe_id
          type: string
        deleted_at:
          format: date-time
          title: deleted_at
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      required:
      - created_at
      - id
      - probe_id
      - updated_at
      title: Probe
    ErrorResponse:
      example:
        message: message
      properties:
        message:
          title: message
          type: string
      required:
      - message
      title: ErrorResponse